Translate exclusion lists to binary scaling factors (Toon).
The force evaluation should no longer depend directly on the exclusion
lists, but rather take this information from the neighborlists. This
change is done in a few steps and it makes easier to access all 1-4
information in the nonbonded code. Instead of binary exclusion (0 or 1)
I introduce a scaling factor from 0 to 1. For now these factors are
just 0 or 1, but that will change in later patches. Interactions that
are scaled to zero are excluded. This information is only stored for
1-2, 1-3 and 1-4 interactions. This will be used to take into account
the EI_SCALE14 and VDW_SCALE14 options. We may add similar EI_SCALE13
etc. options in future.